Improve omelette page styling mobile, responsivity
I improved the omelette page's styling on mobile, providing a larger
image that fits the whole screen width on different browsers. Additiona-
lly, I removed redundancies in the CSS, improved maintanability, as well
as optimized my HTML. I had to add two images for them to change betw-
een mobile and desktop view but added an aria-hidden attribute to the
image not contained in <main> as to improve accessibility for people
using screen readers.